webserver/requirements.txt
Flask==3.0.2
gunicorn==21.2.0
Pillow==10.3.0